refactor: refactoring of extensions and code-gen
