+++ packages/sealed_currencies/lib/sealed_currencies.dart
@@ -1,7 +1,9 @@
 /// Provides data for world currencies in the form of sealed classes.
 library sealed_currencies;
 
+export "package:sealed_languages/sealed_languages.dart";
+
 export "src/data/fiat_currencies.data.dart";
-export "src/helpers/fiat_currency/fiat_currency_formatting_extension.dart";
-export "src/helpers/fiat_currency/fiat_currency_getters_extension.dart";
+export "src/helpers/extensions/fiat_currency/fiat_currency_format.dart";
+export "src/helpers/extensions/fiat_currency/fiat_currency_getters.dart";
 export "src/model/currency.dart";
