+++ packages/sealed_currencies/test/src/helpers/fiat_currency/fiat_currency_getters_extension_test.dart
@@ -1,8 +1,8 @@
-import "package:sealed_currencies/src/helpers/fiat_currency/fiat_currency_getters_extension.dart";
+import "package:sealed_currencies/src/helpers/extensions/fiat_currency/fiat_currency_getters.dart";
 import "package:sealed_currencies/src/model/currency.dart";
 import "package:test/test.dart";
 
-void main() => group("FiatCurrencyFormattingExtension", () {
+void main() => group("FiatCurrencyGetters", () {
       final value = FiatCurrency.list.last;
 
       test("iso4217Letter", () => expect(value.iso4217Letter, value.code));
